Rename project: when a rename will change the project's name field, but not the actual disk name, abort the front-end rename procedure, and add a safeguard to the back-end to prevent deleting the source file when it's the same as the target file. Ref http://bugs.activestate.com/show_bug.cgi?id=94803 rn=
(integrated from komodo (trunk) change 72623 by ericp)

git-svn-id: http://svn.openkomodo.com/repos/openkomodo/trunk@10513 b87d16e7-d29f-4ee1-b5ec-817d645d335f

+++ src/projects/koProjectService.py
@@ -297,11 +297,16 @@ class KoPartService(object):
                 newContents = m.group(1) + m.group(2) + newName + m.group(2) + m.group(3)
             else:
                 log.error("Can't find any name attribute in file %s", oldPath)
+        if sys.platform.startswith("linux"):
+            isSameBasename = newName == oldName
+        else:
+            isSameBasename = newName.lower() == oldName.lower()
         if newContents:
             fd = open(newPath, 'wb')
             fd.write(newContents)
             fd.close()
-            os.unlink(oldPath)
-        else:
+            if not isSameBasename:
+                os.unlink(oldPath)
+        elif not isSameBasename:
             os.rename(oldPath, newPath)
